Avoid the extra layer of providers now it can be handled within cascading values
